+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changes since 1.1.2
 - feat: add wait option
+- feat: apiVersion 1.1 for wait option and add better version handling
 
 # 1.1.2
 - chore: dependency upgrades